requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-pptx==0.6.21
lxml>=4.3.0
Pillow>=3.3.0
requests>=2.31.0
